Change bitfields from enum to uint64 (#132)
Implements spec change https://github.com/open-telemetry/opamp-spec/pull/125

In some languages enum fields are strongly typed and it is impossible to assigned OR-ed values to the enum field. This makes impossible to compose the bit fields properly.

This changes all bit field declarations from enum to uint64. The enum declarations stay, so that bit definitions are clear.

+++ client/internal/clientcommon.go
@@ -214,7 +214,7 @@ func (c *ClientCommon) PrepareFirstMessage(ctx context.Context) error {
 			msg.EffectiveConfig = cfg
 			msg.RemoteConfigStatus = c.ClientSyncedState.RemoteConfigStatus()
 			msg.PackageStatuses = c.ClientSyncedState.PackageStatuses()
-			msg.Capabilities = c.Capabilities
+			msg.Capabilities = uint64(c.Capabilities)
 		},
 	)
 	return nil
